# speed up repeat()