This PR makes `repeat()` much faster when `nbits` is not a large multiple of 8.  It can provide speedups around 100x in many cases, and around 1000x for 1 or 2 bits.

It is tangentially related to https://github.com/ilanschnell/bitarray/issues/132.

 The current implementation of `repeat()` is already very efficient when`nbits` is a large multiple of 8; This optimization makes it as efficient in many more cases by doubling the number of bits copied at once until it is a large multiple of 8.  This makes it possible for `copy_n()` to efficiently use `memmove()` instead of copying each bit individually.

When `nbits` is a large multiple of 8, the performance is essentially the same as the current version.



Here are some timings with and without the optimization: 

```python
%timeit bitarray('1' * 7) * 10 ** 8
1.11 s ± 2.82 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)       # original version
7.39 ms ± 42 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)     # optimized version

%timeit bitarray('10') * 10 ** 8
453 ms ± 1.55 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)       # original version
610 µs ± 25.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)   # optimized version

%timeit bitarray('1') * 10 ** 8
322 ms ± 540 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)        # original version
187 µs ± 1.07 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # optimized version

%timeit bitarray('1' * 8) * 10 ** 8
520 ms ± 11.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)       # original version
9.26 ms ± 50.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   # optimized version  

# current implementation is already fast in this case
%timeit bitarray('1' * 8) * (10 ** 4) * (10 ** 4)  
8.57 ms ± 31.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   # original version
8.71 ms ± 27.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   # optimized version

```