@ilanschnell 2021-07-17
Thank you for this PR!  I think it can be simplified.  Why the `(bits_to_copy % 8 != 0 || bits_to_copy < 10000)` part?  One can simply keep doubling until half the final size is reached and then copy the remaining part.  Something like:
```
    Py_ssize_t k = nbits;
    while (k <= n * nbits / 2) {
        copy_n(self, k, self, 0, k);
        k *= 2;
    }
    copy_n(self, k, self, 0, nbits * n - k);
```