@danielg1111 2021-07-17
Yes, it can be simplified, although I suspect the simplification would make it somewhat slower (at least it did when I tried something similar).  My theory is that the upper limit on the size tends to improve performance by keeping it small enough to fit in cache.