@ilanschnell 2021-07-18
Ah, thank you for the explanation.  I didn't think about the cache size.  I'll merge it now, but probably do the simplification (after some measurements).  The speedup will still be very significant, and I'd rather not bake assumptions (like the cache size) into the code.  Thanks again!