@ilanschnell 2021-07-18
I remove the "cache optimization" for the sake of simplicity.  On my machine:
```
%timeit bitarray('1') * 10 ** 9
24.9 ms ± 119 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # without cache optimization
24.1 ms ± 128 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # this PR
4.96 s ± 154 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # original version
```