@danielg1111 2021-07-18
Interesting.  Here's the same benchmark on my machine with and without the attempt at cache optimization:
```
%timeit bitarray('1') * 10 ** 9
16.1 ms ± 75.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # without cache optimization
11.3 ms ± 37.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # this PR 
```
I figured 10,000 bits would be small enough to fit in cache on basically any computer and have similar results, but I guess not.  Or the speedup could have been from something other than caching.