@ilanschnell 2021-07-18
It also depend on how the string library function `memmove()` is implemented, because this is where all the work is happening.

It also just occurred to me that one could do:
```
    while (k <= self->nbits / 2) {  /* double copies */
        if (k == 8) {
            char c = self->ob_item[0];
            memset(self->ob_item, c, (size_t) Py_SIZE(self));
            return 0;
        }
        copy_n(self, k, self, 0, k);
        k *= 2;
    }
```
That is, when we're 8 bits use `memset()`.  However, this will only bring a speedup for bitarrays of size `1`, `2`, `4` and `8`, and the speedup is not that much:
```
%timeit bitarray('1') * 10 ** 9
20.4 ms ± 184 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # with memset
24.9 ms ± 119 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # current version
```